api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
torch==2.5.1
transformers==4.46.3
safetensors==0.4.5
langchain-groq==0.2.1
langchain-core==0.3.21
python-dotenv==1.0.1